% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ncolumn.R
\name{Ncolumn}
\alias{Ncolumn}
\title{Forma de nombrar columnas ya definidas}
\usage{
Ncolumn(data, columna, nuevo_nombre)
}
\arguments{
\item{nuevo_nombre}{}
}
\value{
a table with names
}
\description{
Forma de nombrar columnas ya definidas
}
\examples{
Ncolumn(genotype_analysis, "X1", "ID")


}
