% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotgenotype.R
\name{Plotgenotype}
\alias{Plotgenotype}
\title{Barplot de genotipos especficos}
\usage{
Plotgenotype(datos, columna)
}
\arguments{
\item{columna}{}
}
\value{
a barplot of specific genotype
}
\description{
Barplot de genotipos especficos
}
\examples{
Plotgenotype(genotype_analysis, "X7")


}
