% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tablecont.R
\name{Tablecont}
\alias{Tablecont}
\title{Conteo de elementos y señalarlos en filas}
\usage{
Tablecont(data, columna)
}
\arguments{
\item{columna}{}
}
\description{
Conteo de elementos y señalarlos en filas
}
\examples{
Tablecont(genotype_analysis, "X2")


}
